requirements.txt
fastapi 
uvicorn 
firebase-admin 
pydantic 
rnumpy 
requests 
python-multipart
python-dotenv
firebase_admin
pillow
